Refactor: Split SyncSection.tsx into smaller components
Refactors the `SyncSection.tsx` file into smaller, more manageable components to improve code organization and readability. This includes extracting sections for active syncs, smart update, and individual sync cards. Unused imports and files will be removed.
