+++ src/components/admin/SyncSection.tsx
@@ -2,22 +2,20 @@
 import { useState, useEffect } from "react";
 import { useQuery } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
-import { SyncCard } from "./sync-card";
 import { SyncType } from "@/types/sync";
 import { ActiveSyncs } from "./sync/ActiveSyncs";
 import { useSyncOperations } from "./hooks/useSyncOperations";
 import { useSyncProgressSubscription } from "./hooks/useSyncProgressSubscription";
 import { SyncProgressDB } from "./types/sync-types";
-import { useToast } from "@/hooks/use-toast";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { WandSparkles } from "lucide-react";
+import { CountryInputSection } from "./sync/CountryInputSection";
+import { SmartUpdateSection } from "./sync/SmartUpdateSection";
+import { SyncCardsGrid } from "./sync/SyncCardsGrid";
+import { useWebhookHandler } from "./sync/useWebhookHandler";
 
 export const SyncSection = () => {
-  const { toast } = useToast();
   const [countryInput, setCountryInput] = useState<string>("");
-  const [smartUpdateBatchSize, setSmartUpdateBatchSize] = useState<number>(25);
   const { isInitializing, clearData, setClearData, handleSync } = useSyncOperations();
+  const { isWebhookLoading, handlePetPoliciesWebhook } = useWebhookHandler();
   
   // Set up sync progress subscription
   useSyncProgressSubscription();
@@ -68,60 +66,17 @@ export const SyncSection = () => {
   });
 
   // Handle smart update for pet policies
-  const handleSmartUpdate = () => {
-    if (smartUpdateBatchSize <= 0) {
-      toast({
-        variant: "destructive",
-        title: "Invalid Batch Size",
-        description: "Please enter a positive number for the batch size.",
-      });
-      return;
-    }
-    
+  const handleSmartUpdate = (batchSize: number) => {
     handleSync('petPolicies', false, 'update', { 
       smartUpdate: true,
-      batchSize: smartUpdateBatchSize
+      batchSize: batchSize
     });
   };
 
-  // Function to handle Pet Policies sync with n8n webhook
-  const handlePetPoliciesWebhook = () => {
-    const webhookUrl = "https://petjumper.app.n8n.cloud/webhook-test/50b1e027-2372-4cf1-95f4-b99f3d700d1e";
-    
-    console.log('Triggering Pet Policies sync via n8n webhook');
-    
-    // Get the current clearData state for pet policies
-    const shouldClearData = clearData.petPolicies;
-    
-    fetch(webhookUrl, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        type: 'pet_policies_sync',
-        clearData: shouldClearData,
-        timestamp: new Date().toISOString()
-      }),
-    })
-    .then(response => {
-      if (!response.ok) {
-        throw new Error(`HTTP error! Status: ${response.status}`);
-      }
-      toast({
-        title: "Pet Policies Sync Initiated",
-        description: "The sync process has been triggered via external service.",
-      });
-      console.log('Webhook response:', response);
-    })
-    .catch(error => {
-      console.error('Error calling webhook:', error);
-      toast({
-        variant: "destructive",
-        title: "Error Starting Sync",
-        description: "Failed to trigger the sync process. The webhook service may be unavailable or misconfigured.",
-      });
-    });
+  // Combine isInitializing with webhook loading state
+  const combinedLoadingState = {
+    ...isInitializing,
+    petPolicies: isInitializing.petPolicies || isWebhookLoading
   };
 
   return (
@@ -129,87 +84,27 @@ export const SyncSection = () => {
       <ActiveSyncs syncProgress={syncProgress || {}} />
 
       {/* Country input for country policies */}
-      <div className="mb-8">
-        <input
-          type="text"
-          placeholder="Enter country name for single country sync"
-          value={countryInput}
-          onChange={(e) => setCountryInput(e.target.value)}
-          className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-        />
-      </div>
+      <CountryInputSection 
+        initialValue={countryInput}
+        onChange={setCountryInput}
+      />
 
       {/* Smart Update section for pet policies */}
-      <div className="mb-8 border p-4 rounded-lg space-y-4">
-        <h3 className="text-lg font-semibold">Smart Update Pet Policies</h3>
-        <div className="grid grid-cols-1 md:grid-cols-4 gap-4 items-center">
-          <div className="md:col-span-3">
-            <div className="flex items-center gap-2">
-              <span className="text-sm whitespace-nowrap">Batch Size:</span>
-              <Input
-                type="number"
-                min="1"
-                max="100"
-                value={smartUpdateBatchSize}
-                onChange={(e) => setSmartUpdateBatchSize(parseInt(e.target.value) || 25)}
-                className="w-24"
-              />
-              <span className="text-xs text-muted-foreground">
-                (Recommended: 25-50 airlines per batch)
-              </span>
-            </div>
-          </div>
-          <Button 
-            variant="outline" 
-            className="border-dashed border-2 h-full"
-            onClick={handleSmartUpdate}
-            disabled={isInitializing.petPolicies}
-          >
-            <WandSparkles className="mr-2" />
-            Run Smart Update
-          </Button>
-        </div>
-        <p className="text-xs text-muted-foreground">
-          Smart Update analyzes airline websites and prioritizes those with pet policies that need updating.
-          The improved algorithm identifies specific pet policy URLs and ensures more accurate updates.
-        </p>
-      </div>
+      <SmartUpdateSection 
+        onSmartUpdate={handleSmartUpdate}
+        isLoading={combinedLoadingState.petPolicies}
+      />
 
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
-        {Object.entries(SyncType).map(([key, value]) => (
-          <SyncCard
-            key={key}
-            title={`${key.replace(/([A-Z])/g, ' $1').trim()} Synchronization`}
-            clearData={clearData[key as keyof typeof SyncType]}
-            onClearDataChange={(checked) => {
-              setClearData(prev => ({ ...prev, [key]: checked }));
-            }}
-            isLoading={isInitializing[value]}
-            onSync={(resume, mode) => {
-              if (key === 'countryPolicies' && mode !== 'clear') {
-                // Only validate country input for single country sync
-                const trimmedCountry = countryInput.trim();
-                if (!trimmedCountry) {
-                  toast({
-                    variant: "destructive",
-                    title: "Country Required",
-                    description: "Please enter a valid country name for single country sync.",
-                  });
-                  return;
-                }
-                handleSync(key as keyof typeof SyncType, resume, trimmedCountry);
-              } else if (key === 'petPolicies') {
-                // Always call webhook for pet policies, let the external process handle resuming
-                handlePetPoliciesWebhook();
-              } else {
-                // For full sync or other types, no country validation needed
-                handleSync(key as keyof typeof SyncType, resume, mode);
-              }
-            }}
-            syncProgress={syncProgress?.[value]}
-          />
-        ))}
-      </div>
+      {/* Sync cards grid */}
+      <SyncCardsGrid 
+        clearData={clearData}
+        setClearData={setClearData}
+        isInitializing={combinedLoadingState}
+        syncProgress={syncProgress}
+        handleSync={handleSync}
+        handlePetPoliciesWebhook={() => handlePetPoliciesWebhook(clearData.petPolicies)}
+        countryInput={countryInput}
+      />
     </div>
   );
 };
